requirements.txt
# --- Core Network & Environment ---
requests==2.33.1
python-dotenv==1.2.2

# --- Image Processing & Vision ---
Pillow==12.2.0
pytesseract==0.3.13

# --- System & GUI Automation ---
psutil==7.2.2

# --- AI & Search Tools ---
ddgs==9.12.1
yfinance==1.2.0

# --- Document Parsing (Advanced) ---
pypdf==5.4.0
html2text==2025.4.15

# --- Voice & Audio ---
gTTS==2.5.4
SpeechRecognition==3.15.2

# --- REST API ---
fastapi==0.135.3
uvicorn[standard]==0.42.0

# --- Setup Utilities ---
setuptools==69.0.3
pyyaml==6.0.3
watchdog==6.0.0
pybreaker==1.4.1
numpy==1.26.4
docker==7.1.0

# --- Database (PostgreSQL + pgvector) ---
psycopg[binary]==3.3.3
psycopg_pool==3.3.0

# --- Observability (OpenTelemetry + Jaeger) ---
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-instrumentation-fastapi==0.61b0
